refactor TestRunner & NativeTestRunner location
